Rework the admin for User and Profile and other stuff
All the management of Profile is done inside the User management in the
Admin view.
I added a verbose name to most of the models field.
The admin views are now clean.

Rework on the activation of User. It is now done on a pre_save signal
and isn't mixed with save_user_profile signal.

Reworked the list_profile view. It now work with the user and not the
profile. It also ignore inactive users.