Add the support for a leisure field
The leisure are define as their own model so that different can share
the same leisure and the link can be done easily.

A form field using select2 is also implemented to allow user to select
their leisure either in the list of already selected leisure by users,
or by entering a new one.